refactor : jpa는 command에서만 사용하므로, 저장구조 변경
